feat: Implement auth

+++ app/api/categories.py
@@ -30,29 +30,51 @@ def read_category(auth_user, id):
 
 
 @app_api.route('/categories/', methods=['POST'])
-def create_category():
+@auth_required()
+def create_category(auth_user):
+    
     category = Category.from_json(request.json)
+    category.user_id = auth_user.get('id')
+
     db.session.add(category)
     db.session.commit()
-    payload = category.to_json()
-    return jsonify(result(payload)), 201, {'Location': url_for('api.read_category', id=category.id)}
+
+    location = url_for('api.read_category', id=category.id)
+    payload = result(category.to_json())
+
+    return jsonify(payload), 201, {'Location': location}
 
 
 @app_api.route('/categories/<int:id>', methods=['PUT'])
-def update_category(id):
-    category = Category.query.get_or_404(id)
+@auth_required()
+def update_category(auth_user, id):
+    user_id = auth_user.get('id')
+    category = Category.query.filter_by(id=id, user_id=user_id)
+
+    if category is None:
+        return page_not_found()
+
     category.name = request.json.get('category', category.name)
     category.type = request.json.get('type', category.type)
     category.go = request.json.get('go', category.go)
+    
     db.session.add(category)
     db.session.commit()
     payload = category.to_json()
+    
     return jsonify(result(payload))
 
 
 @app_api.route('/categories/<int:id>', methods=['DELETE'])
+@auth_required()
 def delete_category(id):
-    category = Category.query.get_or_404(id)
+    
+    user_id = auth_user.get('id')
+    category = Category.query.filter_by(id=id, user_id=user_id)
+
+    if category is None:
+        return page_not_found()
+    
     db.session.delete(category)
     db.session.commit()
     return jsonify(result(payload={})), 202
